fixed issue

+++ src/main/java/com/example/libms/author/AuthorServiceImpl.java
@@ -1,6 +1,8 @@
 package com.example.libms.author;
 
+import com.example.libms.author.exceptions.AuthorHasBooksException;
 import com.example.libms.author.exceptions.AuthorNotFoundException;
+import com.example.libms.book.BookRepository;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,11 +12,13 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class AuthorService {
-    private static final Logger log = LoggerFactory.getLogger(AuthorService.class);
+public class AuthorServiceImpl implements IAuthorService {
+    private static final Logger log = LoggerFactory.getLogger(AuthorServiceImpl.class);
     private final AuthorRepository repository;
+    private final BookRepository bookRepository;
     private final AuthorMapper mapper;
 
+    @Override
     public List<AuthorDto> getAllAuthors() {
         return repository.findAll()
                 .stream()
@@ -22,6 +26,7 @@ public class AuthorService {
                 .toList();
     }
 
+    @Override
     public AuthorDto getAuthorById(Long id) {
         Author author = repository.findById(id)
                 .orElseThrow(() -> new AuthorNotFoundException("no author with id: " + id));
@@ -29,21 +34,29 @@ public class AuthorService {
         return mapper.toDto(author);
     }
 
+    @Override
     public AuthorDto createAuthor(AuthorDto dto) {
         Author savedAuthor = repository.save(mapper.toEntity(dto));
         log.info("added new author {}", savedAuthor.getName());
         return mapper.toDto(savedAuthor);
     }
 
+    @Override
     public void deleteAuthor(Long id) {
         if (!repository.existsById(id)) {
             throw new AuthorNotFoundException("no author with id: " + id);
+        }
+
+        int authorBooksCount = bookRepository.countAuthorBooks(id);
+        if (authorBooksCount > 0) {
+            throw new AuthorHasBooksException("author with id " + id + " has " + authorBooksCount + " books");
         }
 
         repository.deleteById(id);
         log.info("Deleted user with id {}", id);
     }
 
+    @Override
     public AuthorDto updateAuthor(Long id, AuthorDto dto) {
         Author existingAuthor = repository.findById(id)
                 .orElseThrow(() -> new AuthorNotFoundException("no author with id " + id));
